refatorando o nome das classe e variaveis e Completar os Eventos para o Rabbit
