+++ pom.xml
@@ -93,6 +93,13 @@
 			<artifactId>jackson-datatype-jsr310</artifactId>
 			<version>2.13.0</version>
 		</dependency>
+		
+		<dependency>
+		    <groupId>org.springframework.boot</groupId>
+		    <artifactId>spring-boot-devtools</artifactId>
+		    <optional>true</optional>
+		    <scope>runtime</scope>
+		</dependency>
 	</dependencies>
 
 	<build>
