+++ src/main/java/net/atos/api/cliente/ApiClienteApplication.java
@@ -1,0 +1,13 @@
+package net.atos.api.cliente;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApiClienteApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ApiClienteApplication.class, args);
+	}
+
+}
