+++ src/main/java/net/atos/api/cliente/domain/StatusPessoaEnum.java
@@ -1,0 +1,6 @@
+package net.atos.api.cliente.domain;
+
+public enum StatusPessoaEnum {
+	ATIVO,
+	INATIVO
+}
