+++ src/main/java/net/atos/api/cliente/domain/TipoContatoEnum.java
@@ -1,0 +1,10 @@
+package net.atos.api.cliente.domain;
+
+public enum TipoContatoEnum {
+	
+	CELULAR,
+	RESIDENCIAL,
+	COMERCIAL,
+	OUTROS
+
+}
