+++ src/main/java/net/atos/api/cliente/domain/TipoPessoaEnum.java
@@ -1,0 +1,7 @@
+package net.atos.api.cliente.domain;
+
+public enum TipoPessoaEnum {
+	
+	FISICA,
+	JURIDICA
+}
