+++ src/main/java/net/atos/api/cliente/events/PessoaDeletaEvent.java
@@ -1,0 +1,17 @@
+package net.atos.api.cliente.events;
+
+import net.atos.api.cliente.domain.PessoaVO;
+
+public class PessoaDeletaEvent {
+	
+	private PessoaVO pessoa;
+	
+	public PessoaDeletaEvent(PessoaVO pPessoa) {
+		this.pessoa = pPessoa;		
+	}
+
+	public PessoaVO getPessoa() {
+		return pessoa;
+	}
+	
+}
