+++ src/main/java/net/atos/api/cliente/repository/PessoaRepository.java
@@ -1,0 +1,19 @@
+package net.atos.api.cliente.repository;
+
+import java.time.LocalDate;
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.repository.PagingAndSortingRepository;
+import org.springframework.stereotype.Repository;
+
+import net.atos.api.cliente.repository.entity.PessoaEntity;
+
+@Repository
+public interface PessoaRepository extends PagingAndSortingRepository<PessoaEntity, Long>{
+	
+	public Page<PessoaEntity> findByDataCadastroBetween(LocalDate dataInicio, LocalDate dataFim, Pageable pageable);
+	
+}
