+++ src/main/java/net/atos/api/cliente/service/CriaPessoa.java
@@ -1,0 +1,14 @@
+package net.atos.api.cliente.service;
+
+import javax.validation.constraints.NotNull;
+
+import net.atos.api.cliente.domain.PessoaVO;
+import net.atos.api.cliente.domain.StatusPessoaEnum;
+import net.atos.api.cliente.domain.TipoPessoaEnum;
+
+public interface CriaPessoa{
+	
+	public PessoaVO persistir(@NotNull(message = "Cliente não pode ser nulo") PessoaVO pessoaVO); 
+
+	public Boolean isType(TipoPessoaEnum type);
+}
