--- src/main/java/net/atos/cliente/ApiClienteApplication.java
@@ -1,13 +1,0 @@
-package net.atos.cliente;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApiClienteApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApiClienteApplication.class, args);
-	}
-
-}
