--- src/main/java/net/atos/cliente/domain/StatusEnum.java
@@ -1,6 +1,0 @@
-package net.atos.cliente.domain;
-
-public enum StatusEnum {
-	ATIVO,
-	INATIVO;
-}
