--- src/main/java/net/atos/cliente/domain/TipoContatoEnum.java
@@ -1,10 +1,0 @@
-package net.atos.cliente.domain;
-
-public enum TipoContatoEnum {
-	
-	CELULAR,
-	RESIDENCIAL,
-	COMERCIAL,
-	OUTROS
-
-}
