--- src/main/java/net/atos/cliente/domain/TipoPessoaEnum.java
@@ -1,7 +1,0 @@
-package net.atos.cliente.domain;
-
-public enum TipoPessoaEnum {
-	
-	FISICA,
-	JURIDICA
-}
