--- src/main/java/net/atos/cliente/service/ClienteInativoEvent.java
@@ -1,17 +1,0 @@
-package net.atos.cliente.service;
-
-import net.atos.cliente.domain.PessoaVO;
-
-public class ClienteInativoEvent {
-
-	private PessoaVO cliente;
-	
-	public ClienteInativoEvent(PessoaVO pcliente) {
-		this.cliente = pcliente;		
-	}
-
-	public PessoaVO getCliente() {
-		return cliente;
-	}
-	
-}
